+++ scripts/build-release.ps1
@@ -1,5 +1,8 @@
 Set-Location "$PSScriptRoot\.."
 dotnet build -c Release
 if ($?) {
-	Copy-Item ".\bin\Release\netstandard2.1\NameplateTweaks.dll" "..\..\BepInEx\plugins"
+	$name = (Get-Content manifest.json | ConvertFrom-Json).name
+	Copy-Item ".\bin\Release\netstandard2.1\$name.dll" "..\..\BepInEx\plugins"
+	$cfg = "..\..\BepInEx\config\BepInEx.cfg"
+	(Get-Content $cfg) -replace "^HideManagerGameObject = true","HideManagerGameObject = false" | Set-Content $cfg
 }