+++ scripts/build.ps1
@@ -1,6 +1,9 @@
 Set-Location "$PSScriptRoot\.."
 dotnet build
 if ($?) {
-	Remove-Item "..\..\BepInEx\plugins\NameplateTweaks.dll" -ErrorAction SilentlyContinue
-	Copy-Item ".\bin\Debug\netstandard2.1\NameplateTweaks.dll" "..\..\BepInEx\scripts"
+	$name = (Get-Content manifest.json | ConvertFrom-Json).name
+	Remove-Item "..\..\BepInEx\plugins\$name.dll" -ErrorAction SilentlyContinue
+	Copy-Item ".\bin\Debug\netstandard2.1\$name.dll" "..\..\BepInEx\scripts"
+	$cfg = "..\..\BepInEx\config\BepInEx.cfg"
+	(Get-Content $cfg) -replace "^HideManagerGameObject = false","HideManagerGameObject = true" | Set-Content $cfg
 }