+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.0.3
+- Fixes an issue where you could occassionally see your own nameplate.
+- Improves the voice activity indicator on the spectator screen
 # 1.0.2
 Fixes [Issue #1](https://github.com/taffyko/LCNameplateTweaks/issues/1)
 # 1.0.1